Update to new error helper signatures in n-api
Update to use new signatures
Update copies of n-api files so they match. I've not done
a direct copy because more work is needed given that the
n-api files now use node internals.  I have updated so that
they are consistent and will build ok with respect to
this specific chaange. I took this approach to minimize
the time the wrapper is broken by the breaking change
on the n-api side.

PR-URL: https://github.com/nodejs/node-addon-api/pull/78
Reviewed-By: Kyle Farnung <kfarnung@microsoft.com>