test: run interfering tests in their own process
Tests such as addon and addon_data both use SetInstanceData. Thus, they
overwrite each other's instance data. We change them both to run in a
child process such that they do not call SetInstanceData on init, but
rather they return a JS function on init which, when called, produces
the actual binding which needs SetInstanceData.

We also introduce a utility function for launching a test in its own
child process for the purpose of running tests that would otherwise
interfere with each other.

Signed-off-by: Gabriel Schulhof <gabrielschulhof@gmail.com>
PR-URL: https://github.com/nodejs/node-addon-api/pull/1325
Reviewed-By: Chengzhong Wu <legendecas@gmail.com>